- Remove exposition of the tkn from the Sandbox - Make loginCallback method private - Implement RedirectionSafari as an NSObject and move it from Sandbox to IdentityCoreSDK
