+++ Sandbox/Sandbox/Base.lproj/Main.storyboard
@@ -88,21 +88,6 @@
             </objects>
             <point key="canvasLocation" x="-259" y="865"/>
         </scene>
-        <!--Redirection View Controller-->
-        <scene sceneID="UuE-Q7-sX8">
-            <objects>
-                <viewController id="ddb-9N-K4B" customClass="RedirectionViewController" customModule="Sandbox" customModuleProvider="target" sceneMemberID="viewController">
-                    <view key="view" contentMode="scaleToFill" id="WzA-bY-qOk">
-                        <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
-                        <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
-                        <color key="backgroundColor" systemColor="systemBackgroundColor" cocoaTouchSystemColor="whiteColor"/>
-                        <viewLayoutGuide key="safeArea" id="4dP-Wf-vvT"/>
-                    </view>
-                </viewController>
-                <placeholder placeholderIdentifier="IBFirstResponder" id="ge5-ns-cEm" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
-            </objects>
-            <point key="canvasLocation" x="-2532" y="791"/>
-        </scene>
         <!--Passwordless-->
         <scene sceneID="pPL-xw-jtO">
             <objects>
